Task2(B)_DroneDetection/requirements.txt
fastapi
uvicorn[standard]
ultralytics
pillow
opencv-python
numpy
python-multipart
